Angular changes

+++ Entities/Category.cs
@@ -7,7 +7,7 @@ namespace Entities;
 
 public partial class Category
 {
-    public short CategoryId { get; set; }
+    public int CategoryId { get; set; }
 
     public string CategoryName { get; set; }
